feat: move test module to commands
