+++ src/commands/mod.rs
@@ -5,3 +5,4 @@ pub mod completions;
 pub mod lint;
 pub mod new;
 pub mod run;
+pub mod test;
