+++ src/lib.rs
@@ -4,6 +4,5 @@ pub(crate) mod bin_target;
 pub mod commands;
 pub(crate) mod config;
 pub(crate) mod external_cli;
-pub mod test;
 #[cfg(feature = "web")]
 pub(crate) mod web;
